UPDATE Files
In Board, we have add the pattern of the board, with that we can see
where are all of the pawns when we start a game

In CelticChess we had modified the fonction startGame, we initialize all
of the parameters in CelticChess now.

In player we have start to add a modification for the color of each
player and the moment when they start to play with a boolean.